# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/admin/payments/reconciliation/scan:
    post:
      operationId: ReconciliationController_scan
      parameters: []
      responses:
        '200':
          description: Scans the database for reconciliation discrepancies and upserts
            issues.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationScanEnvelopeDto'
        '401': &a49
          description: Missing, invalid, or expired Supabase JWT.
        '403': &a50
          description: Admin role is required.
      security: &a51
        - supabaseAuth: []
      tags: &a52
        - admin-reconciliation
  /v1/admin/payments/reconciliation/issues:
    get:
      operationId: ReconciliationController_listIssues
      parameters: []
      responses:
        '200':
          description: Reconciliation issues scoped by campus for campus admins.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationIssueListEnvelopeDto'
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
  /v1/admin/payments/reconciliation/issues/{issueId}:
    get:
      operationId: ReconciliationController_getIssue
      parameters:
        - name: issueId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Reconciliation issue detail with investigation notes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationIssueDetailEnvelopeDto'
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
  /v1/admin/payments/reconciliation/issues/{issueId}/verify-payment:
    post:
      operationId: ReconciliationController_verifyPayment
      parameters:
        - name: issueId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Re-verifies the issue payment against Paystack.
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
  /v1/admin/payments/reconciliation/issues/{issueId}/retry-webhook:
    post:
      operationId: ReconciliationController_retryWebhook
      parameters:
        - name: issueId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Retries a failed webhook by re-verifying its payment.
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
  /v1/admin/payments/reconciliation/issues/{issueId}/review:
    post:
      operationId: ReconciliationController_reviewIssue
      parameters:
        - name: issueId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Marks a reconciliation issue investigating/resolved/ignored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationIssueEnvelopeDto'
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
  /v1/admin/payments/reconciliation/issues/{issueId}/notes:
    post:
      operationId: ReconciliationController_addNote
      parameters:
        - name: issueId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Adds an internal investigation note to a reconciliation issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationNoteEnvelopeDto'
        '401': *a49
        '403': *a50
      security: *a51
      tags: *a52
...
info:
  title: Meal Direct API
  description: Stable REST API for Meal Direct web and future mobile clients.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    supabaseAuth:
      scheme: bearer
      bearerFormat: Supabase JWT
      type: http
      description: Supabase Auth access token.
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for important mutations such as order creation and payments.
    requestId:
      type: apiKey
      in: header
      name: x-request-id
      description: Optional caller-provided request ID. The API echoes or generates one.
    operationsToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Internal operations token. Temporary until Super Admin JWT/RBAC is
        implemented.
  schemas:
...
    ReconciliationScanResultDto:
      type: object
      properties:
        detected:
          type: number
      required:
        - detected
    ReconciliationScanEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationScanResultDto'
      required:
        - data
    ReconciliationIssueRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        issueType:
          type: string
          enum: &a98
            - initialized_unconfirmed
            - paid_order_pending
            - webhook_processing_failed
            - provider_success_not_local
            - duplicate_success
            - amount_mismatch
            - currency_mismatch
            - refund_stuck
        status:
          type: string
          enum: &a99
            - open
            - investigating
            - resolved
            - ignored
        severity:
          type: string
          enum: &a100
            - info
            - warning
            - critical
        paymentId:
          type: string
          nullable: true
        orderId:
          type: string
          nullable: true
        refundId:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        providerReference:
          type: string
          nullable: true
        detail:
          type: object
        firstDetectedAt:
          type: string
        lastDetectedAt:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        resolutionNote:
          type: string
          nullable: true
      required:
        - id
        - issueType
        - status
        - severity
        - detail
        - firstDetectedAt
        - lastDetectedAt
    ReconciliationIssueListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReconciliationIssueRecordDto'
      required:
        - data
    ReconciliationNoteRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        issueId:
          type: string
          format: uuid
        authorId:
          type: string
          nullable: true
        body:
          type: string
        createdAt:
          type: string
      required:
        - id
        - issueId
        - body
        - createdAt
    ReconciliationIssueDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        issueType:
          type: string
          enum: *a98
        status:
          type: string
          enum: *a99
        severity:
          type: string
          enum: *a100
        paymentId:
          type: string
          nullable: true
        orderId:
          type: string
          nullable: true
        refundId:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        providerReference:
          type: string
          nullable: true
        detail:
          type: object
        firstDetectedAt:
          type: string
        lastDetectedAt:
          type: string
        reviewedBy:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        resolutionNote:
          type: string
          nullable: true
        notes:
          type: array
          items:
            $ref: '#/components/schemas/ReconciliationNoteRecordDto'
      required:
        - id
        - issueType
        - status
        - severity
        - detail
        - firstDetectedAt
        - lastDetectedAt
        - notes
    ReconciliationIssueDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationIssueDetailDto'
      required:
        - data
    ReconciliationIssueEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationIssueRecordDto'
      required:
        - data
    ReconciliationNoteEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationNoteRecordDto'
      required:
        - data
...